VM/values/types: [getValuePair] cast `bool`s to `byte` directly (`CastSizes` warning)

+++ src/vm/values/types.nim
@@ -364,8 +364,8 @@ template getValuePair*(): untyped =
 
     (cast[uint32](ord(xKind)) shl 16.uint32) or 
     (cast[uint32](ord(yKind))) or  
-    (cast[uint32](cast[uint32](xKind==Integer) * cast[uint32](x.iKind==BigInteger)) shl 31) or
-    (cast[uint32](cast[uint32](yKind==Integer) * cast[uint32](y.iKind==BigInteger)) shl 15)
+    (cast[uint32](cast[byte](xKind==Integer) * cast[byte](x.iKind==BigInteger)) shl 31) or
+    (cast[uint32](cast[byte](yKind==Integer) * cast[byte](y.iKind==BigInteger)) shl 15)
 
 proc `||`*(va: static[ValueKind | IntegerKind], vb: static[ValueKind | IntegerKind]): uint32 {.compileTime.}=
     ## generate a ValuePair value for given va and vb Value kinds
